+++ ax25.h
@@ -96,7 +96,7 @@ extern struct ax_route Ax_default;
 
 /* Link quality report packet header, internal format */
 struct lqhdr {
-	uint16 version;		/* Version number of protocol */
+	uint version;		/* Version number of protocol */
 #define	LINKVERS	1
 	int32	ip_addr;	/* Sending station's IP address */
 };
@@ -125,10 +125,10 @@ struct lq {
 	/* # packets reported as transmitted by station as of his last update */
 	int32 lasttxcnt;
 
-	uint16 hisqual;	/* Fraction (0-1000) of station's packets heard
+	uint hisqual;	/* Fraction (0-1000) of station's packets heard
 			 * as of last update
 			 */
-	uint16 myqual;	/* Fraction (0-1000) of our packets heard by station
+	uint myqual;	/* Fraction (0-1000) of our packets heard by station
 			 * as of last update
 			 */
 #endif
@@ -154,7 +154,7 @@ struct ax_route *ax_lookup(uint8 *);
 void ax_recv(struct iface *,struct mbuf **);
 int axui_send(struct mbuf **bp,struct iface *iface,int32 gateway,uint8 tos);
 int axi_send(struct mbuf **bp,struct iface *iface,int32 gateway,uint8 tos);
-int ax_output(struct iface *iface,uint8 *dest,uint8 *source,uint16 pid,
+int ax_output(struct iface *iface,uint8 *dest,uint8 *source,uint pid,
 	struct mbuf **data);
 int axsend(struct iface *iface,uint8 *dest,uint8 *source,
 	int cmdrsp,int ctl,struct mbuf **data);
@@ -169,7 +169,7 @@ void getlqhdr(struct lqhdr *hp,struct mbuf **bpp);
 void logsrc(struct iface *iface,uint8 *addr);
 void logdest(struct iface *iface,uint8 *addr);
 char *putlqentry(char *cp,uint8 *addr,int32 count);
-char *putlqhdr(char *cp,uint16 version,int32 ip_addr);
+char *putlqhdr(char *cp,uint version,int32 ip_addr);
 struct lq *al_lookup(struct iface *ifp,uint8 *addr,int sort);
 
 /* In ax25subr.c: */
